requirements.txt
langchain_openai==0.1.16
langchain_experimental==0.0.62
langchainhub==0.1.20
langchain==0.2.7
openai==1.35.13
numpy==1.26.4
streamlit==1.36.0
tqdm==4.66.4
ipykernel==6.29.5
PyYAML==6.0.1
streamlit-chat==0.1.1
pinecone-client==4.1.2